+++ app/(tabs)/SleepFacts.jsx
@@ -1,26 +1,26 @@
 import React from 'react';
-import { View, Text, Image, FlatList, StyleSheet } from 'react-native';
-
+import { View, Text, Image, FlatList, StyleSheet, TouchableOpacity } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
 const facts = [
   {
     id: '1',
     text: 'Newborns sleep up to 16-17 hours a day, but in short bursts of 2-4 hours.',
-    image: require('../assets/images/bby.png'),  
+    image: require('../../assets/images/bby.png'),  
   },
   {
     id: '2',
     text: 'Blue light exposure suppresses melatonin, making it harder to fall asleep.',
-    image: require('../assets/images/light.png'), 
+    image: require('../../assets/images/light.png'), 
   },
   {
     id: '3',
     text: 'Regular meditation reduces insomnia and can lead to longer, deeper sleep.',
-    image: require('../assets/images/meditation.png'),
+    image: require('../../assets/images/meditation.png'),
   },
   {
     id: '4',
     text: 'Power naps lasting 10-20 minutes can improve alertness and performance.',
-    image: require('../assets/images/nap.png'), 
+    image: require('../../assets/images/nap.png'), 
   },
 ];
 
@@ -31,9 +31,12 @@ const SleepFacts = () => {
       <Text style={styles.factText}>{item.text}</Text>
     </View>
   );
-
+  const navigation = useNavigation();
   return (
     <View style={styles.container}>
+      <TouchableOpacity onPress={() => navigation.navigate('sleepTrack')}>
+                <Image source={require('../../assets/back-icon.png')} className="w-8 h-8 mt-12" />
+              </TouchableOpacity>
       <Text style={styles.header}>Interesting Facts</Text>
       <FlatList
         data={facts}
@@ -49,12 +52,14 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     padding: 20,
-    backgroundColor: '#fff',
+    backgroundColor: '#ffff',
+    marginTop: 50,
   },
   header: {
     fontSize: 24,
     fontWeight: 'bold',
     marginBottom: 20,
+    textAlign: 'center',
   },
   listContainer: {
     paddingBottom: 20,
