+++ app/(tabs)/home.jsx
@@ -6,9 +6,9 @@ import { useNavigation } from 'expo-router';
 const Home = () => {
   const navigation = useNavigation();
 
-  
+
   const icons = {
-    mealPlanning: require('../../assets/images/meal.png'), 
+    mealPlanning: require('../../assets/images/meal.png'),
     symptomLogging: require('../../assets/images/image.png'),
     sleepTracking: require('../../assets/images/sleep.png'),
     healthGoalSetting: require('../../assets/images/goalsetting.png'),
@@ -17,116 +17,112 @@ const Home = () => {
   return (
     <SafeAreaView className="bg-white h-full">
       <ScrollView >
-      <View
+        <View
           className="w-full flex-1 justify-center h-full px-7 my-6"
           style={{
             minHeight: Dimensions.get("window").height - 100,
           }}
         >
-          {/* Greeting */}
-          <Text className="text-3xl font-bold text-black mb-6 text-center">Wellness Pioneers</Text>
-
-          {/* Tile Buttons */}
           <View className="space-y-4">
             {/* Meal Planning */}
             <TouchableOpacity
-    className="bg-[#FFE3CB]  rounded-lg p-4 flex-row items-center"
-    style={{
-      shadowColor: '#000',    // Shadow color
-      shadowOffset: { width: 0, height: 4 },  // Shadow offset
-      shadowOpacity: 0.3,     // Shadow opacity
-      shadowRadius: 4,        // Shadow radius
-      elevation: 8,           // Shadow for Android
-    }}
-    onPress={() => navigation.navigate('meal_planner')}
-  >
-    <Image 
-      source={icons.mealPlanning} 
-      style={{ 
-        width: 70,            
-        height: 70,           
-        marginRight: 20 
-      }} 
-    />
-    <Text className="text-xl font-bold text-black">   {/* Larger font size */}
-      Meal Planning
-    </Text>
-  </TouchableOpacity>
+              className="bg-[#FFE3CB]  rounded-lg p-4 flex-row items-center"
+              style={{
+                shadowColor: '#000',    // Shadow color
+                shadowOffset: { width: 0, height: 4 },  // Shadow offset
+                shadowOpacity: 0.3,     // Shadow opacity
+                shadowRadius: 4,        // Shadow radius
+                elevation: 8,           // Shadow for Android
+              }}
+              onPress={() => navigation.navigate('meal_planner')}
+            >
+              <Image
+                source={icons.mealPlanning}
+                style={{
+                  width: 70,
+                  height: 70,
+                  marginRight: 20
+                }}
+              />
+              <Text className="text-xl font-bold text-black">   {/* Larger font size */}
+                Meal Planning
+              </Text>
+            </TouchableOpacity>
 
             {/* Symptom Logging */}
-  <TouchableOpacity
-    className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
-    style={{
-      shadowColor: '#000',
-      shadowOffset: { width: 0, height: 4 },
-      shadowOpacity: 0.3,
-      shadowRadius: 4,
-      elevation: 8,
-    }}
-    onPress={() => navigation.navigate('symptomsList')}
-  >
-    <Image 
-      source={icons.symptomLogging} 
-      style={{ 
-        width: 70,       
-        height: 70,      
-        marginRight: 20 
-      }} 
-    />
-    <Text className="text-xl font-bold text-black">  {/* Larger font size */}
-      Symptom Logging
-    </Text>
-  </TouchableOpacity>
+            <TouchableOpacity
+              className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
+              style={{
+                shadowColor: '#000',
+                shadowOffset: { width: 0, height: 4 },
+                shadowOpacity: 0.3,
+                shadowRadius: 4,
+                elevation: 8,
+              }}
+              onPress={() => navigation.navigate('symptomsList')}
+            >
+              <Image
+                source={icons.symptomLogging}
+                style={{
+                  width: 70,
+                  height: 70,
+                  marginRight: 20
+                }}
+              />
+              <Text className="text-xl font-bold text-black">  {/* Larger font size */}
+                Symptom Logging
+              </Text>
+            </TouchableOpacity>
 
-  {/* Sleep Tracking */}
-  <TouchableOpacity
-    className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
-    style={{
-      shadowColor: '#000',
-      shadowOffset: { width: 0, height: 4 },
-      shadowOpacity: 0.3,
-      shadowRadius: 4,
-      elevation: 8,
-    }}
-    onPress={() => navigation.navigate('sleepTrack')}
-  >
-    <Image 
-      source={icons.sleepTracking} 
-      style={{ 
-        width: 70,        
-        height: 70,       
-        marginRight: 20 
-      }} 
-    />
-    <Text className="text-xl font-bold text-black">   {/* Larger font size */}
-      Sleep Tracking
-    </Text>
-  </TouchableOpacity>
+            {/* Sleep Tracking */}
+            <TouchableOpacity
+              className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
+              style={{
+                shadowColor: '#000',
+                shadowOffset: { width: 0, height: 4 },
+                shadowOpacity: 0.3,
+                shadowRadius: 4,
+                elevation: 8,
+              }}
+              onPress={() => navigation.navigate('sleepTrack')}
+            >
+              <Image
+                source={icons.sleepTracking}
+                style={{
+                  width: 70,
+                  height: 70,
+                  marginRight: 20
+                }}
+              />
+              <Text className="text-xl font-bold text-black">   {/* Larger font size */}
+                Sleep Tracking
+              </Text>
+            </TouchableOpacity>
 
-  {/* Health Goal Setting */}
-  <TouchableOpacity
-    className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
-    style={{
-      shadowColor: '#000',
-      shadowOffset: { width: 0, height: 4 },
-      shadowOpacity: 0.3,
-      shadowRadius: 4,
-      elevation: 8,
-    }}
-    onPress={() => navigation.navigate('nutrition_add')}
-  >
-    <Image 
-      source={icons.healthGoalSetting} 
-      style={{ 
-        width: 70,        
-        height: 70,       
-        marginRight: 20 
-      }} 
-    />
-    <Text className="text-xl font-bold text-black">  {/* Larger font size */}
-      Health Goal Setting
-    </Text>
-  </TouchableOpacity>
+            {/* Health Goal Setting */}
+            <TouchableOpacity
+              className="bg-[#FFE3CB] rounded-lg p-4 flex-row items-center"
+              style={{
+                shadowColor: '#000',
+                shadowOffset: { width: 0, height: 4 },
+                shadowOpacity: 0.3,
+                shadowRadius: 4,
+                elevation: 8,
+              }}
+              onPress={() => navigation.navigate('nutrition_add')}
+            >
+              <Image
+                source={icons.healthGoalSetting}
+                style={{
+                  width: 70,
+                  height: 70,
+                  marginRight: 20
+                }}
+              />
+              <Text className="text-xl font-bold text-black">  {/* Larger font size */}
+                Health Goal Setting
+              </Text>
+            </TouchableOpacity>
           </View>
         </View>
       </ScrollView>
